requirements.txt
annotated-types==0.7.0
anyio==4.7.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.8
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
distlib==0.3.9
distro==1.9.0
exceptiongroup==1.2.2
filelock==3.16.1
fonttools==4.51.0
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
graphic==1.0.0a6
graphviz==0.20.3
grpcio==1.68.1
grpcio-status==1.68.1
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
jiter==0.8.0
joblib==1.2.0
kiwisolver==1.4.5
matplotlib==3.8.4
neo4j==5.25.0
nltk==3.9.1
numpy==2.2.5
openai==1.57.0
packaging==24.0
pandas==2.2.3
patsy==0.5.6
pillow==10.3.0
platformdirs==4.3.6
proto-plus==1.25.0
protobuf==5.29.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.3
pydantic_core==2.27.1
pygame==2.5.2
pyparsing==3.0.9
PyQt6==6.7.1
PyQt6-Qt6==6.7.3
PyQt6_sip==13.9.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyWavelets==1.4.1
regex==2024.11.6
requests==2.32.3
rsa==4.9
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
six==1.17.0
sklearn==0.0.post1
sniffio==1.3.1
statsmodels==0.14.2
threadpoolctl==3.1.0
tifffile==2022.10.10
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.2.3
virtualenv==20.28.0
xgboost==2.0.3
